+++ RandomizerScreen/Randomizer.tsx
@@ -1,9 +1,7 @@
 import { StyleSheet, Text, View, ScrollView } from 'react-native';
 import React, { useState, ReactElement } from 'react';
-import neighborhoodInfo from './NeighborhoodInfo';
-import googlePlacesFetch from './GoogleAPIFetch';
-import PlaceItem from './PlaceItem';
-import SpinButton from './SpinButton';
+import PlaceItem from '../PlaceItem';
+import SpinButton from '../SpinButton';
 
 export interface Place {
   name: string;
@@ -26,9 +24,10 @@ export default function Randomizer(): ReactElement {
         horizontal={true}
       >
         <View style={styles.resultsContainer}>
-          {nearbyRestaurants.map((restaurant, i) => (
-            <PlaceItem key={i} place={restaurant} />
-          ))}
+          {nearbyRestaurants &&
+            nearbyRestaurants.map((restaurant, i) => (
+              <PlaceItem key={i} place={restaurant} />
+            ))}
         </View>
       </ScrollView>
       <Text style={styles.neighborhood}>{spinResult.name}</Text>
@@ -56,14 +55,9 @@ const styles = StyleSheet.create({
     fontSize: 50,
     padding: 20,
   },
-  scrollViewContainer: {
-    // alignItems: 'center',
-    // width: '100%',
-    // backgroundColor: 'transparent',
-  },
+  scrollViewContainer: {},
   searchResults: {
     display: 'flex',
-    //borderWidth: 1,
     maxHeight: '70%',
   },
   stars: {
